package restructuring and refactor of the FileManager class
