+++ src/main/java/org/aksw/simba/squirrel/analyzer/compress/enums/MimeTypeEnum.java
